django_backend/requirements.txt
Django>=5.0,<7
djangorestframework>=3.15
django-cors-headers>=4.4
Pillow>=10.0
djangorestframework-camel-case>=1.4.0
djangorestframework-simplejwt>=5.3
twilio>=9.0
python-dotenv>=1.0
gunicorn>=22.0
whitenoise>=6.7
